[test] Fix deadlock in AvfIntegrationTest
This cl resolves the deadlock issue in issue in
AvfIntegrationTest that arises from the timeout
handling of the AndroidJUnit4 test runner. The
presence of the timeout causes the test setup and
the test method to be executed on different threads,
resulting in the test becoming stuck.

To fix this, this cl uses the BlockJUnit4ClassRunner
and JUnit's Timeout rule to ensure that the test
setup and test method are executed on the same thread.

Bug: 335627002
Test: atest AvfRkpdAppIntegrationTests
Change-Id: I42833f6a45bb8da84db2a35c50878137dbed139a