aconfigd: simplify and repurpose aconfigd_socket_test
Before this change, aconfigd_socket_test is locking down the actual
aconfigd behavior (how it handles different requests). We now moved to
use aconfigd_test with mocks for that purposes (so that we can test the
aconfigd request handling logic even if aconfigd is not turned on device).
Now the purpose of aconfigd_socket_test is to ensure we can properly
connect and communicate to aconfigd socket when aconfigd is enabled.
The test points are guarded by the flag and won't run if the flag value
is false.

Bug: b/312444587
Test: atest aconfigd_test
Change-Id: I9dfd5c186341726132d6e78a8e28ff8ec058b56b