No need to read max number of EE at NFC_ENABLE_REVT
Maximum number of EE is defined as default to NFA_EE_MAX_EE_SUPPORTED.
It is then updated with the value of configuration
NAME_NFA_MAX_EE_SUPPORTED when the JNI is initialized.
There is no need to read it again when NFC_ENABLE_REVT is received,
as it is only read again from the the configuration file.

Bug: 370360038
Change-Id: I45c3e08a399484d9a558797f19bc11aea37ca82f