Part 1: Fix finalize_sdk script execution for "all modules" case
aosp/3005020 broke the finalize_sdk script for the special case when
no modules are passed as parameter.
Though it works correctly when any non-empty list of subset of valid
mainline modules is passed.

Emtpy module list implies all modules are finalized. In that case, all
module_sdk zip files are fetched and is extracted by finding the
corresponding prebuilts repo for the module_sdk.

Currently this mapping is from module_name to its prebuilts repo.
This fix adds sdk_name to the mainline-modules-info.json file which
help to find the prebuilts repo using the sdk_name of the module.

This fix is split into 2 CLs -
1. Adds sdk_name to mainline-modules-info.json.
2. Update finalize_sdk script to use sdk_name to fetch prebuilts repo.

This CL addressed part #1.

Bug: 321970705
Test: atest --host mainline_modules_sdks_test --no-bazel-mode
Change-Id: I83ec8a994f851a484be25df405e4b99e1b7392fb