Autoupdate of 2 submodules
*art updated to 00e6586769b75c84d3b5bd33e96faf4dac0bdf45
Add ART MTS eng-only tests to the test plan of ART MTS shard 03.

This is a quick and easy way to run ART MTS eng-only tests as part of
ART test code coverage configurations (so that they can contribute to
the line coverage of the tested code), without requiring the creation
of a new ATP test shard.

Restructuring the MTS generation logic to make it cleaner is planned
as a follow-up.

Flag: TEST_ONLY
Test: m mts && mts-tradefed run commandAndExit mts-art
Test: m mts && mts-tradefed run commandAndExit mts-art-shard-03
Bug: 268491069
Bug: 167385698
Change-Id: I7249b8e3ffe1be411ba08c2e74218a050d570ff8

*test/mts updated to 23b9f01c829cf8bf7f515259a329d095492fe339
Merge changes from topics "art-mts-eng-only-tests-in-shard-03-test-plan", "regen-test-files-2024-06-07" into main

* changes:
  Regenerate the ART MTS definition (2024-06-07 (b)).
  Regenerate the ART MTS definition (2024-06-07 (a)).